+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Consider downgrading Python requirement (at least to 3.10, but I dont know how low it can gowhere)
 
+### Added
+- Show actual cost - as calculated by langchain - after call. Note: does not seem to work for gpt-4o
+
 ## [0.8.3]
 
 ### Fixed